ok ACTUALLY THIS TIME bc had to fix gallery particles re-rendering
